isOnGround method
Simple is the best

for isOnGround, we ONLY need to know if player is TOUCHING the wall.

Overlap and Cast catch just every objects inside their sight, so it is not that proper approach.

But, IsTouching (with LayerMask) only checks if player rigidbody is touching something specific, even Effector.

Only one line, intuitive and neat code. Couldn't be better!